### 2021-05-13 #3279 @whedon
[PRE REVIEW]: The Region Growing Plugin (RegionGrow)
**Submitting author:** @gro5-AberUni (<a href="http://orcid.org/0000-0002-5054-1548">Gregory Oakes</a>)
**Repository:** <a href="https://github.com/gro5-AberUni/RegionGrow" target ="_blank">https://github.com/gro5-AberUni/RegionGrow</a>
**Version:** V4.1
