### 2021-05-28 #3279 @gro5-AberUni
@hugoledoux Thank you for including the newer citation, I will update the paper to reflect that.

The main difference between the plugin we present here and the Semi Automatic Classification Plugin is that the region growing tool in the Semi Automatic Classification plugin is integrated within the wider toolbox with the expectation that digitised regions are used for image classification within the plugin itself. We present a more general tool for digitising features through region growing where the resultant features are used in other workflows. We also found that the Semi automatic classification plugin can be quite difficult to use, particularly for non-experts who do not frequently use GIS software so we aimed to create a tool with an easy to use UI.
