+++ paper/references.bib
@@ -8,15 +8,17 @@
   year={2012}
 }
 
-@article{congedo2016semi,
-  title={Semi-automatic classification plugin documentation},
-  author={Congedo, Luca},
-  journal={Release},
-  volume={4},
-  number={0.1},
-  pages={29},
-  year={2016}
+@misc{Cogendo2021,
+author = {Cogendo, L.},
+doi = {10.5281/ZENODO.4587408},
+keywords = {GIS,image analysis,image processing,land cover,raster analysis,remote sensing,supervised classification,surface reflectance},
+month = {mar},
+pages = {208},
+title = {Semi-Automatic Classification Plugin},
+url = {https://zenodo.org/record/4587408},
+year = {2021}
 }
+
 
 @inproceedings{niu2014image,
   title={Image segmentation algorithm for disease detection of wheat leaves},
