+++ paper/references.bib
@@ -9,12 +9,12 @@
 }
 
 @misc{Cogendo2021,
+title = {Semi-Automatic Classification Plugin},
 author = {Cogendo, L.},
 doi = {10.5281/ZENODO.4587408},
 keywords = {GIS,image analysis,image processing,land cover,raster analysis,remote sensing,supervised classification,surface reflectance},
 month = {mar},
 pages = {208},
-title = {Semi-Automatic Classification Plugin},
 url = {https://zenodo.org/record/4587408},
 year = {2021}
 }
