+++ paper/references.bib
@@ -9,8 +9,9 @@
 }
 
 @misc{Cogendo2021,
-title = {Semi-Automatic Classification Plugin},
+title = {Semi-Automatic Classification Plugin (Version v7.8)},
 author = {Cogendo, L.},
+journal= {Zenodo}
 doi = {10.5281/ZENODO.4587408},
 keywords = {GIS,image analysis,image processing,land cover,raster analysis,remote sensing,supervised classification,surface reflectance},
 month = {mar},
