+++ paper/references.bib
@@ -11,7 +11,7 @@
 @misc{Cogendo2021,
 title = {Semi-Automatic Classification Plugin (Version v7.8)},
 author = {Cogendo, L.},
-journal= {Zenodo}
+journal = {Zenodo},
 doi = {10.5281/ZENODO.4587408},
 keywords = {GIS,image analysis,image processing,land cover,raster analysis,remote sensing,supervised classification,surface reflectance},
 month = {mar},
