+++ paper/references.bib
@@ -15,7 +15,6 @@ journal = {Zenodo},
 doi = {10.5281/ZENODO.4587408},
 keywords = {GIS,image analysis,image processing,land cover,raster analysis,remote sensing,supervised classification,surface reflectance},
 month = {mar},
-pages = {208},
 url = {https://zenodo.org/record/4587408},
 year = {2021}
 }
