### 2021-06-17 #3346 @Geogoeroe
Hi @gro5-AberUni,
I got stuck before actually really trying the plugin.... I used the description on downloading, but... The plugin manager says it doesn’t install:
![image](https://user-images.githubusercontent.com/9982061/122407315-d408a800-cf81-11eb-819d-505ba93e12ce.png)

However, it does… with that __MACOSX as an invalid extra plugin.

So I gave it a go anyway. The plugin loads data (both from a UAV image and from a Sentinel dataset).

When the grow function is used (start) it unerringly comes up with the message to increase the treshold. After trying several values in the range from 0-15000, I considered that the plugin probably wasn't installed in a decent way.
![image](https://user-images.githubusercontent.com/9982061/122407743-26e25f80-cf82-11eb-8a9b-935cbe5de141.png)

I'm running QGIS 3.18 on a Windows 10 environment.
Any help would be appreciated.

Cheers,
Erik.
