### 2021-06-18 #3346 @gro5-AberUni
> Hi @gro5-AberUni,
> I got stuck before actually really trying the plugin.... I used the description on downloading, but... The plugin manager says it doesn’t install:
> ![image](https://user-images.githubusercontent.com/9982061/122407315-d408a800-cf81-11eb-819d-505ba93e12ce.png)
>
> However, it does… with that __MACOSX as an invalid extra plugin.
>
> So I gave it a go anyway. The plugin loads data (both from a UAV image and from a Sentinel dataset).
>
> When the grow function is used (start) it unerringly comes up with the message to increase the treshold. After trying several values in the range from 0-15000, I considered that the plugin probably wasn't installed in a decent way.
> ![image](https://user-images.githubusercontent.com/9982061/122407743-26e25f80-cf82-11eb-8a9b-935cbe5de141.png)
>
> I'm running QGIS 3.18 on a Windows 10 environment.
> Any help would be appreciated.
>
> Cheers,
> Erik.

Hi Erik, Thanks for getting in touch.

The first error you received: "No named '__MACOSX/regiongrow'" shouldn't cause any issues, the ZIP directory that QGIS uses to install the plugin was created on a MACOS system which automatically includes a directory '__MACOSX' containing data and resource forks for the zip directory which QGIS was attempting to install as a plugin. I've adjusted the Zip directory within the GitHub repo to not include that directory so you shouldn't see that error.

The Second issue you have noted I believe was caused by a bug where if you digitise a feature close to the left edge of the image, where the neighbourhood would extend beyond the boundaries of the image the plugin would use an incorrect raster origin value and could not digitise the features correctly.
I have addressed this issue and tested the plugin on a Windows 10 Machine running QGIS 3.18.3 as well as a MacOS and linux environment and it seems to work correctly.

Please let me know if you come across any further issues
Thanks
Greg
