### 2021-06-19 #3346 @jorgegil
@gro5-AberUni  Would it be possible to share the different image files that you use in the paper and documentation?

That would allow reproducing exactly what is presented. It can be cropped images just of the sections shown in the paper. And these could be included in the repository.

Thanks
Jorge
