### 2021-06-21 #3346 @gro5-AberUni
> @gro5-AberUni Would it be possible to share the different image files that you use in the paper and documentation?
>
> That would allow reproducing exactly what is presented. It can be cropped images just of the sections shown in the paper. And these could be included in the repository.
>
> Thanks
> Jorge

Hi @jorgegil, I have uploaded some example UAV imagery that was used in the documentation and the figures in the paper to the location: gro5-AberUni/RegionGrow/Example_UAV_Data/ so you should be able to reproduce the results we present.

Thanks
Greg
