### 2021-06-21 #3346 @Geogoeroe
Ah, thanks for the test data. However, on de-installing and re-installing the new version of the plugin, I got this:
![image](https://user-images.githubusercontent.com/9982061/122726939-cc454e00-d276-11eb-8d1d-47da991b77c2.png)

Another thing is, that if I deactivate the plugin, it doesn't clear the menu-item on the plugin menu. So after activiating the plugin again (three times), I get this:
![image](https://user-images.githubusercontent.com/9982061/122727233-1cbcab80-d277-11eb-87c0-10cf6fff7060.png)
