### 2021-06-21 #3346 @Geogoeroe
Ah, indeed, that was the issue with the data... I didn't interpret the python error as something data-related, but that took care of it. However, I still have to check the manual properly (I might be doing something wrong here), as the default settings still come up with the treshold warning:
2021-06-21T11:36:03     WARNING    Region Grower Plugin : Not Enough Spectral Difference Found. Increase the Threshold
2021-06-21T11:36:03     SUCCESS    Region Grower Plugin : Process Successful
