### 2021-06-21 #3346 @Geogoeroe
For the updating of the plugins: Yes, I noticed that the menu cleans after restarting QGIS, it is not a permanent thing. However, other plugins don't show this specific behaviour... So it might be a simple trick somewhere.
