### 2021-06-29 #3346 @jorgegil
> For the updating of the plugins: Yes, I noticed that the menu cleans after restarting QGIS, it is not a permanent thing. However, other plugins don't show this specific behaviour... So it might be a simple trick somewhere.

Indeed, @gro5-AberUni you must use signals to add and remove the plugin from the menu and the toolbar when the plugin is loaded / unloaded. This happens in the main class.

It's rare behaviour (people do not check/uncheck a plugin multiple times) but it has a dev solution and other plugins do it correctly.
