### 2021-06-30 #3346 @jorgegil
I've completed a review of the software and the paper and have a long list of comments for both. I'll add them here as separate comments, grouped where relevant, to allow response.

General comment on the paper for now:
- The paper is well written but in my view lacks technical detail to make the algorithm clearer and more transparent (details to follow).
- A more general use case for the tool must be explained, or indicated that this is meant for the specific use only described in projects.
- It lacks a comparison with an existing "Magic Wand" plugin for QGIS.
- The images need better captions and be referenced in the text, otherwise they fall out of place and don't help. The paper lacks references.
