### 2021-06-30 #3346 @jorgegil
Minor tweaks to the paper:
- paragraph 1: "at a continental or global scape" should be scale
- paragraph 2: "RegionGrow generates connects pixels with a similar colour" generates should be dropped
