### 2021-06-30 #3346 @jorgegil
Paper figures:
- figure 1 seems out of place. Either add descriptions to the caption (a-f) as in later figures or refer to it in the text. The steps in "Using this software" have no relation to the steps in figure 1
- in fact, you need to refer to all images in the text so we look at them in context and they support the descriptions
