### 2021-06-30 #3346 @jorgegil
Use case for the tool:
- the section heading "ongoing projects" is not immediately obvious what it refers to: software projects, or applications of the tool.
- Maybe it should be a section describing scope, uses cases, and limitations, to explain how it is meant to be used in practice, but also in research. What what it cannot do.
- For a tool called region grow I was expecting something like a magic wand in Photoshop: threshold is colour tolerance only, there's no neighbourhood size, and the user can choose contiguous regions or multiple regions. There is a plugin for QGIS called.. "Magic Wand". How is this one different or better/more appropriate for specific tasks?
