### 2021-06-30 #3346 @jorgegil
Questions on Step 2 in the paper:
- Why does the user need to indicate a distance? Why 25 as a default?
- Text should mention "neighbourhood size (m)" which is the label in the tool. The description must be consistent with the text in the GUI.
- Why a square neighbourhood, not circle or hexagon?
