### 2021-06-30 #3346 @jorgegil
Questions on Step 3 in the paper:
- What kind of 'spatial weighting' is it? linear? How is it calculated, in metres?
- Shouldn't the spatial relation be based on contiguity rather than distance?
