### 2021-06-30 #3346 @jorgegil
Questions on Step 4 in the paper:
- At this stage an equation with all aspects of steps 3 and 4, even if simple, would be valuable to make the algorithm very transparent.
- Are the values used scaled prior to the sum? What is their range?
- What name do you give the 'combined value'?
- Needs an explanation for the user defined threshold. How are users supposed to know the right value? Are there reference values (from theory, empirical) or is it trial and error (to get the feel for it)? Why is 15 the default value?
- This has an impact on the workflow and its reliability/reproducibility. I see there's more information in the technical guide, this should be included in the paper.
