### 2021-06-30 #3346 @jorgegil
Questions on Step 5 in the paper:
- Why give option for buffer? It should be explained why the option is there, and example of when/how to use.
- Why only geoJSON or shapefile output? Could also create a memory layer like many other plugins and let the user choose how to save it.
