### 2021-06-30 #3346 @jorgegil
Testing and validation questions:
- Was there any testing and validation done to see if the tool selects what is intended? Should there be a test data set with correct regions to understand thresholds?
- Unless the results are not meant to be precise and only rough indications, which applies to a specific use case. In that sense the tool is not a general region identification tool.
- Could two rather different colour values that are very close spatially be added to a region, while similar distant ones and still contiguous not? If in the use case for the tool this is not critical, it should be indicated explicitly so that the tool is not misused in other cases.
- For a general tool called region grow I was expecting something like a magic wand in Photoshop: threshold is colour tolerance only, there's no neighbourhood size, and the user can choose contiguous regions or multiple regions. There is already such a plugin for QGIS called.. "Magic Wand". How is Region Grow one different or better/more appropriate for specific tasks?
