### 2021-06-30 #3346 @jorgegil
Tool usability:
- starting state of image type radio buttons. Having a default, should it not be drone?
- Image type could be a drop down menu, with Drone imagery on top (the intended use).
- What is "Band combinations" for? I don't see explanation in the paper, just a brief mention in the technical guide.
