### 2021-06-30 #3346 @jorgegil
Tool output:
- Difference between output file and existing vector data set unclear and not explained. Having two options is confusing.
- In the layers panel, the added vector layer should have the name given in the "Output vector filename" box, currently has no name.
- Existing vector layer could give the option to choose from an existing vector data set from the layers panel?
- vector output: why saved with the imagery and  no option to choose another location?
- vector output results in a memory layer: would give flexibility to save in other formats, or postgresql
