### 2021-06-30 #3346 @jorgegil
Tool usability:
- Start button should not be active if basic conditions are not met, like having an output file name, or image type.
- Ideally there should be a toolbar for the red cross circle button, the class and the undo button. Going to the dialogue box next to start and finish for those actions is not intuitive
