### 2021-07-05 #3346 @hugoledoux
@gro5-AberUni the idea is that you consider @jorgegil comments and improve the plugin and let us know here what was modified and how.

👋 @Geogoeroe you haven't ticked any of the chechboxes above, is it because you don't have write access? If so I can reinvite you.
