+++ paper/paper.md
@@ -37,7 +37,7 @@ notoriously time consuming, labour intensive and error prone and, as
 such, the manual digitisation of features in operational projects is
 often unfeasible, particularly in time-sensitive applications like near
 real-time monitoring applications, emergency response mapping or where
-applications are made at a continental or global scape. Consequently,
+applications are made at a continental or global scaple. Consequently,
 there is a need for a simple to use region growing tool that can be used
 to speed up the digitisation process and minimise human error.
 
@@ -45,9 +45,8 @@ We present a new Region Growing plugin tool (RegionGrow) for the open
 source Geographical Information System software QGIS. RegionGrow is
 designed to be user-friendly, computationally efficient, freely
 available, with functionality to accommodate a variety of sources of
-multiband remotely sensed imagery. By selecting a pixel in the centre of
-an object of interest, e.g. a forest, water body etc, RegionGrow
-generates connects pixels with a similar colour based on its Euclidean
+multiband remotely sensed imagery to digitise discrete features and save these features into a vector dataset. By selecting a pixel in the centre of
+an object of interest, e.g. a forest, water body etc, RegionGrow connects pixels with a similar colour based on its Euclidean
 distance with a L\*A\*B transformed colour space. This rapidly speeds up
 the manual digitising process and minimises error.
 
@@ -56,25 +55,28 @@ height="5.369911417322835in"}
 
 # Statement of Need
 
-Within the open source software domain, there is just one region growing tool available from within the Semi-Automatic Classification (SAC) QGIS plugin [@Cogendo2021]. However, this is intended to be used within the SAC classification workflow and cannot be used intuitively for other digitising needs. 
+Within the open source software domain, there are very region growing tools available. Within the Semi-Automatic Classification (SAC) QGIS plugin [@Cogendo2021] is a region growing tool to select like pixels. However, this is intended to be used within the SAC classification workflow and cannot be used intuitively for other digitising needs. In addition there is the ‘Magic Wand’ [@Silvestri2017] plugin which are able to select regions within imagery of similar colour, within a set distance from the point of click and save those to a single vector dataset. However if desired features are outside of this set distance they will not be included in the vector dataset and a new vector layer will have to be produced to include these layers. 
 
-An independent region growing tool is needed for two primary reasons: 1) Where users wish to collect training data for conducting image classifications outside of the SAC plugin tool, for instance, taking advantage of the most current and robust machine learning solutions from within Python libraries like TPOT [@olson2016tpot] or Scikit-Learn [@pedregosa2011scikit]. 2) Where users wish to extract thematic information directly rather than following a classification routine. 
+An independent discrete feature region growing tool is needed for two primary reasons: 1) Where users wish to collect training data for conducting image classifications outside of the SAC plugin tool, for instance, taking advantage of the most current and robust machine learning solutions from within Python libraries like TPOT [@olson2016tpot] or Scikit-Learn [@pedregosa2011scikit]. 2) Where users wish to extract thematic information directly rather than following a classification routine. 
+
+# Scope, Use Case and Limitations
+
+The RegionGrow plugin was developed to allow non-experts to identify and digitise discrete targets of interest within UAV Imagery. This has been deployed operationally with the Spatial Intelligence System (SIS) for precision larviciding based in Zanzibar.
+
+It was then developed to allow for the generation of training datasets which can be used in supervised image classification of a wide range of remotely sensed data, where a single vector dataset is produced. This dataset produces a series of features of differing attributes which relate to the target classes a user is trying to detect. 
+
[…]
