+++ paper/references.bib
@@ -66,3 +66,11 @@ year = {2021}
   year={2011},
   publisher={JMLR. org}
 }
+
+@misc{Silvestri2017,
+author = {Silvestri, Gianluca},
+title = {{Magic Wand}},
+url = {https://plugins.qgis.org/plugins/MagicWand/},
+year = {2017}
+}
+
