+++ paper/paper.md
@@ -116,7 +116,7 @@ of a similar colour.
 ![3D feature space showing pixels within the local area of the
 user clicked location. The cross within the plot marks the user clicked
 pixel.](Figure4.png){width="3.937007874015748in"
-height="2.984346019247594in"}
+height="3.084346019247594in"}
 
 ![a) RGB Drone Imagery, b) L\*A\*B\* transformed drone imagery and
 the location of the user clicked location. C) Euclidean distance within
@@ -128,7 +128,6 @@ A linear spatial weighting is also applied where pixels geographically further
 from the selected pixel have a lower weighting. This is done to ensure that pixels of similar colour but are far away from the users points of click are not considered to be part of the feature the user is trying to digitise.
 
 
-<br/><br/>
 **Step 4.**
 
 Pixels which can be considered to be of the same region can be defined where the colour distance, summed with the sptial distance have a value greater than the user defined threshold as described in equation 1. Where this condition is not met, pixels are set to set to 0  and considered to be no data values. 
