+++ Paper/SubmittedVersion/paper/paper.md
@@ -1,0 +1,165 @@
+---
+title: 'The Region Growing Plugin (RegionGrow)'
+tags:
+  - QGIS
+  - Region Growing
+  - GIS
+  - Remote Sensing
+  - Python
+authors:
+  - name: Gregory Oakes^[Corresponding Author]
+    orchid: 0000-0002-5054-1548
+    affiliation: 1
+  - name: Andy Hardy
+    orchid: 0000-0002-7928-8873
+    affiliation: 1
+  - name: Yussuf Said Yussuf
+    affiliation: "2, 3"
+affiliations:
+  - name: Department of Geography and Earth Sciences, Aberystwyth University, Aberystwyth SY23 3DB, UK
+    index: 1
+  - name: The State University of Zanzibar (SUZA), Tunguu, P.O.Box 146, Zanzibar
+    index: 2
+  - name: Tanzania Flying Labs
+    index: 3
+date: 21-04-2021
+bibliography: references.bib
+
+---
+
+# Summary
+
+In the fields of image analysis, computer vision and remote sensing
+there are a number of instances where the digitisation of features by
+the human operator is needed to develop mapped products or for
+generating training data for classification routines. This process is
+notoriously time consuming, labour intensive and error prone and, as
+such, the manual digitisation of features in operational projects is
+often unfeasible, particularly in time-sensitive applications like near
+real-time monitoring applications, emergency response mapping or where
+applications are made at a continental or global scaple. Consequently,
+there is a need for a simple to use region growing tool that can be used
+to speed up the digitisation process and minimise human error.
+
+We present a new Region Growing plugin tool (RegionGrow) for the open
+source Geographical Information System software QGIS. RegionGrow is
+designed to be user-friendly, computationally efficient, freely
+available, with functionality to accommodate a variety of sources of
+multiband remotely sensed imagery. By selecting a pixel in the centre of
+an object of interest, e.g. a forest, water body etc, RegionGrow connects pixels with a similar colour based on its Euclidean
+distance with a L\*A\*B transformed colour space. This rapidly speeds up
+the manual digitising process and minimises error.
+
+![Overview of the processing steps performed by RegionGrow](Figure1.png){width="3.937007874015748in"
+height="5.369911417322835in"}
+
+# Statement of Need
+
+Within the open source software domain, there is just one region growing tool available from within the Semi-Automatic Classification (SAC) QGIS plugin [@Cogendo2021]. However, this is intended to be used within the SAC classification workflow and cannot be used intuitively for other digitising needs. 
+
+An independent region growing tool is needed for two primary reasons: 1) Where users wish to collect training data for conducting image classifications outside of the SAC plugin tool, for instance, taking advantage of the most current and robust machine learning solutions from within Python libraries like TPOT [@olson2016tpot] or Scikit-Learn [@pedregosa2011scikit]. 2) Where users wish to extract thematic information directly rather than following a classification routine. 
+
+# Ongoing Projects
+
+The RegionGrow QGIS plugin tool was originally developed for use within
+the project: Spatial Intelligence System (SIS) for precision larviciding
+based in Zanzibar, United Republic of Tanzania. Here, UAV technology is
+being used to map pools of water where malarial mosquitoes breed. By
+mapping these targets, they can be located by ground teams and treated
+with low-toxicity larvicide in an effort to reduce the malarial mosquito
+population. Although high resolution images (\<10 cm) of areas of
+interest can be produced rapidly using UAVs, identifying and digitising
+water body features is a laborious and error-strewn process. Using
+RegionGrow, non-GIS specialists can now rapidly and accurately digitise
+these features, supplying vital information to ground-teams
[…]
