### 2021-07-30 0c3c2b3
Delete Paper directory
