+++ Paper/paper/paper.md
@@ -1,164 +1,0 @@
----
-title: 'The Region Growing Plugin (RegionGrow)'
-tags:
-  - QGIS
-  - Region Growing
-  - GIS
-  - Remote Sensing
-  - Python
-authors:
-  - name: Gregory Oakes^[Corresponding Author]
-    orchid: 0000-0002-5054-1548
-    affiliation: 1
-  - name: Andy Hardy
-    orchid: 0000-0002-7928-8873
-    affiliation: 1
-  - name: Yussuf Said Yussuf
-    affiliation: "2, 3"
-affiliations:
-  - name: Department of Geography and Earth Sciences, Aberystwyth University, Aberystwyth SY23 3DB, UK
-    index: 1
-  - name: The State University of Zanzibar (SUZA), Tunguu, P.O.Box 146, Zanzibar
-    index: 2
-  - name: Tanzania Flying Labs
-    index: 3
-date: 21-04-2021
-bibliography: references.bib
-
----
-
-# Summary
-
-In the fields of image analysis, computer vision and remote sensing
-there are a number of instances where the digitisation of features by
-the human operator is needed to develop mapped products or for
-generating training data for classification routines. This process is
-notoriously time consuming, labour intensive and error prone and, as
-such, the manual digitisation of features in operational projects is
-often unfeasible, particularly in time-sensitive applications like near
-real-time monitoring applications, emergency response mapping or where
-applications are made at a continental or global scaple. Consequently,
-there is a need for a simple to use region growing tool that can be used
-to speed up the digitisation process and minimise human error.
-
-We present a new Region Growing plugin tool (RegionGrow) for the open
-source Geographical Information System software QGIS. RegionGrow is
-designed to be user-friendly, computationally efficient, freely
-available, with functionality to accommodate a variety of sources of
-multiband remotely sensed imagery to digitise discrete features and save these features into a vector dataset. By selecting a pixel in the centre of
-an object of interest, e.g. a forest, water body etc, RegionGrow connects pixels with a similar colour based on its Euclidean
-distance with a L\*A\*B transformed colour space. This rapidly speeds up
-the manual digitising process and minimises error.
-
-# Statement of Need
-
-Within the open source software domain, there are very region growing tools available. Within the Semi-Automatic Classification (SAC) QGIS plugin [@Cogendo2021] is a region growing tool to select like pixels. However, this is intended to be used within the SAC classification workflow and cannot be used intuitively for other digitising needs. In addition there is the ‘Magic Wand’ [@Silvestri2017] plugin which are able to select regions within imagery of similar colour, within a set distance from the point of click and save those to a single vector dataset. However if desired features are outside of this set distance they will not be included in the vector dataset and a new vector layer will have to be produced to include these layers. 
-
-An independent discrete feature region growing tool is needed for two primary reasons: 1) Where users wish to collect training data for conducting image classifications outside of the SAC plugin tool, for instance, taking advantage of the most current and robust machine learning solutions from within Python libraries like TPOT [@olson2016tpot] or Scikit-Learn [@pedregosa2011scikit]. 2) Where users wish to extract thematic information directly rather than following a classification routine. 
-
-# Scope, Use Case and Limitations
-
-The RegionGrow plugin was developed to allow non-experts to identify and digitise discrete targets of interest within UAV Imagery. This has been deployed operationally with the Spatial Intelligence System (SIS) for precision larviciding based in Zanzibar.
-
[…]
