+++ Paper/SubmittedVersion/paper/references.bib
@@ -1,68 +1,0 @@
-@article{baldevbhai2012color,
-  title={Color image segmentation for medical images using L* a* b* color space},
-  author={Baldevbhai, Patel Janakkumar and Anand, RS},
-  journal={IOSR Journal of Electronics and Communication Engineering},
-  volume={1},
-  number={2},
-  pages={24--45},
-  year={2012}
-}
-
-@misc{Cogendo2021,
-title = {Semi-Automatic Classification Plugin (Version v7.8)},
-author = {Cogendo, L.},
-journal = {Zenodo},
-doi = {10.5281/ZENODO.4587408},
-keywords = {GIS,image analysis,image processing,land cover,raster analysis,remote sensing,supervised classification,surface reflectance},
-month = {mar},
-url = {https://zenodo.org/record/4587408},
-year = {2021}
-}
-
-
-@inproceedings{niu2014image,
-  title={Image segmentation algorithm for disease detection of wheat leaves},
-  author={Niu, Xiaojing and Wang, Meili and Chen, Xianqiang and Guo, Shihui and Zhang, Hongming and He, Dongjian},
-  booktitle={Proceedings of the 2014 International Conference on Advanced Mechatronic Systems},
-  pages={270--273},
-  year={2014},
-  organization={IEEE}
-}
-
-@software{pandy2017rgblab,
-  title={RGB to CIELab Color Space Conversion.},
-  author={Pandey, Manoj},
-  url={https://gist.github.com/manojpandey/f5ece715132c572c80421febebaf66ae},
-  year={2017},
-  organization={IEEE}
-}
-
-@article{rathore2012colour,
-  title={Colour based image segmentation using L* a* b* colour space based on genetic algorithm},
-  author={Rathore, Vivek Singh and Kumar, Messala Sudhir and Verma, Ashwini},
-  journal={International Journal of Emerging Technology and Advanced Engineering},
-  volume={2},
-  number={6},
-  pages={156--162},
-  year={2012},
-  publisher={Citeseer}
-}
-
-@inproceedings{olson2016tpot,
-  title={TPOT: A tree-based pipeline optimization tool for automating machine learning},
-  author={Olson, Randal S and Moore, Jason H},
-  booktitle={Workshop on automatic machine learning},
-  pages={66--74},
-  year={2016},
-  organization={PMLR}
-}
-
-@article{pedregosa2011scikit,
-  title={Scikit-learn: Machine learning in Python},
-  author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
-  journal={the Journal of machine Learning research},
-  volume={12},
-  pages={2825--2830},
-  year={2011},
-  publisher={JMLR. org}
-}
