+++ region_grow/region_grow.py
@@ -22,7 +22,7 @@
 """
 from qgis.PyQt.QtCore import QSettings, QTranslator, QCoreApplication, QFileInfo
 from qgis.PyQt.QtGui import QIcon
-from qgis.PyQt.QtWidgets import QAction, QMessageBox
+from qgis.PyQt.QtWidgets import QAction, QMessageBox, QLabel, QLineEdit
 from qgis.gui import QgsMapToolEmitPoint, QgsMapTool, QgsMapCanvas,QgsMessageBar,QgsMapToolPan
 
 # Initialize Qt resources from file resources.py
@@ -307,7 +307,7 @@ class RegionGrower:
         callback,
         enabled_flag=True,
         add_to_menu=True,
-        add_to_toolbar=True,
+        add_to_toolbar=False,
         status_tip=None,
         whats_this=None,
         parent=None):
@@ -377,14 +377,125 @@ class RegionGrower:
     def initGui(self):
         """Create the menu entries and toolbar icons inside the QGIS GUI."""
 
+        # Add toolbar
+        self.toolbar = self.iface.addToolBar("RegonGrow")
+
+        self.titleLabel = QLabel()
+
+        self.titleLabel.setText(' RegionGrow  ')
+
+        self.toolbar.addWidget(self.titleLabel)
+
+
+
+        # Add Open Dialog Box to the Toolbar #
+
         icon_path = ':/plugins/region_grow/icon.png'
-        self.add_action(
+        self.openDia = self.add_action(
             icon_path,
             text=self.tr(u'RegionGrow'),
             callback=self.run,
             parent=self.iface.mainWindow())
+        self.toolbar.addAction(self.openDia)
+        # Added #
 
-        # will be set False in run()
+        # Add Undo Button #
+
+            # Add Text Label#
+
+        self.undoLabel = QLabel()
+
+        self.undoLabel.setText(' Undo: ')
+
+        self.toolbar.addWidget(self.undoLabel)
+
+        self.undoTB = QAction(QIcon(':/plugins/region_grow/Undo.png'),
+                               QCoreApplication.translate("MyPlugin", "Undo"),
+                               self.iface.mainWindow())
+
+        # Connect action signals to slots
+        self.undoTB.triggered.connect(self.undo)
+
+        self.undoTB.setEnabled(False)
+
+        self.toolbar.addAction(self.undoTB)
+
+
+        # Add Reactivate Tool #
+
+        self.clickLabel = QLabel()
+
+        self.clickLabel.setText(' Activate Tool: ')
+
+        self.toolbar.addWidget(self.clickLabel)
+
+
+        self.reClickTB = QAction(QIcon(':/plugins/region_grow/clickToolTB.png'),
+                               QCoreApplication.translate("MyPlugin", "Reactivate Tool"),
+                               self.iface.mainWindow())
+
+        # Connect action signals to slots
+        self.reClickTB.triggered.connect(self.activateClickTool)
+
+        self.reClickTB.setEnabled(False)
+
+        self.toolbar.addAction(self.reClickTB)
+
+        # Add Finish Button #
+
+        self.finishLabel = QLabel()
+
+        self.finishLabel.setText(' Finish: ')
+
+        self.toolbar.addWidget(self.finishLabel)
+
+        self.finishTB = QAction(QIcon(':/plugins/region_grow/finished.png'),
+                               QCoreApplication.translate("MyPlugin", "Finish"),
+                               self.iface.mainWindow())
+
+        # Connect action signals to slots
+        self.finishTB.triggered.connect(self.finish)
+
+        self.finishTB.setEnabled(False)
+
+        self.toolbar.addAction(self.finishTB)
+
+        # Add Class Value to Toolbar #
+
+            # Add Text Label#
+
[…]
