### 2021-07-31 #3346 @gro5-AberUni
@jorgegil @hugoledoux

Thank you @jorgegil for you comments and suggestions about the plugin and the paper. I have implemented a series of changes to the plugin specifically within the GUI to improve the user performance, notably I have have added a tool bar containing the main buttons for the operation of the plugin in, as well as the option for the user to adjust the feature class value or buffer distance if desired. I have also adjusted the Main Dialogue box to only use one box to specify the output vector dataset.

I have addressed some of the comments that you have made regarding specific aspects of the paper and the operation of the algorithm in general below, but it has not been possible to include or address all aspects within the paper itself due to the word limitations on the paper.

Questions on Step 2 in the paper:

Why does the user need to indicate a distance? Why 25 as a default?
As the plugin is designed to highlight a discrete feature around the user’s click, the distance outlines how far around this point the plugin should search for similar pixels which are likely to belong to that feature. As only one feature is to be digitised per click it is not necessary to examine every pixel within the image to determine if it is similar. A distance of 25m metres was selected as it allows for large features to be disgusted in one click, while balancing computational efficiency.
Text should mention "neighbourhood size (m)" which is the label in the tool. The description must be consistent with the text in the GUI.
Adjusted in the paper to include Neighbourhood Size (m)
Why a square neighbourhood, not circle or hexagon?
A square neighbourhood was used as it is simpler to implement than a circle or a hexagon neighbourhood and from initial testing a square neighbourhood provides adequate outputs.

Questions on Step 3 in the paper:

What kind of 'spatial weighting' is it? linear? How is it calculated, in metres?
Spatial weighting is linear based on euclidean distance from the centre of the generated neighbourhood outwards

Questions on Step 4 in the paper:

At this stage an equation with all aspects of steps 3 and 4, even if simple, would be valuable to make the algorithm very transparent. An Equation has been added to the paper in step 4 to indicate how colour distance and the spatial weighting is applied to select candidate pixels.
Are the values used scaled prior to the sum? What is their range?
From development and testing it was decided that values dont need to be scaled to be within a specific range
What name do you give the 'combined value'?
In the algorithm it is referred to as totalWeightedDistance but this definition has little relevance to the user.
Needs an explanation for the user defined threshold. How are users supposed to know the right value? Are there reference values (from theory, empirical) or is it trial and error (to get the feel for it)? Why is 15 the default value?
Default values were selected for image types during testing and development and are provided as a guide, in the user guide it will be included that the user can change this based on the general feel of how the tool is performing. The undo feature can be helpful to allow a user to tune the digitised feature.
This has an impact on the workflow and its reliability/reproducibility. I see there's more information in the technical guide, this should be included in the paper.

Questions on Step 5 in the paper:

Why give option for buffer? It should be explained why the option is there, and example of when/how to use.
The option for a buffer on the digitised feature is included to allow the user to select a larger region for example if they are using the tool to highlight an area to direct to field workers. This would be inappropriate if the user was using the tool to generate training data for a supervised image classification as spurious pixels would be included.
Why only geoJSON or shapefile output? Could also create a memory layer like many other plugins and let the user choose how to save it.
GeoJson and and Shapefile formats are used as they are the common formats that have been used within the Spatial Intelligence System for Precision Larviciding where the plugin was initially developed. It is possible to export the digitised layer after processing has finished by exporting the layer in the desired format from the layers panel.

Testing and validation questions:

Was there any testing and validation done to see if the tool selects what is intended? Should there be a test data set with correct regions to understand thresholds?
Testing was performed to ensure that he plugin was digitising features suitably but it was decided that no formal validation routine was particularly necessary particularly with user defined variables.
The above is necessary, unless the results are not meant to be precise and only rough indications of regions, which applies to a specific use case. In that sense the tool is not a general region identification tool.
The results are not designed to a definitive mask layer rather more of an indication, as is the case with other region growing tools such as that sound in the Semi Automatic classification plugin or MagicWand.
Could two rather different colour values that are very close spatially be added to a region, while similar distant ones and still contiguous not? If in the use case for the tool this is not critical, it should be indicated explicitly so that the tool is not misused in other cases.
It would be unlikely to find pixels with very different colours in the same digitised features as when plotted in the 3D colour feature space these pixels will have a very large colour distance which would be greater than the influence of their spatial proximity.
Continuous pixels of similar colour are more likely to be digitised as the same feature as the colour distance within the 3D feature space has a more influence than the spatial weighting
This will be outlined in the paper.

Tool GUI clarity:

Have a starting state of "image type" radio buttons, strange to see all unchecked, not the normal state. Having a default, should it not be Drone?
GUI adjusted to have the drone option as default.
Image type could be a drop down menu, with Drone imagery on top (the intended use).
What is "Band combinations" for? I don't see explanation in the paper, just a brief mention in the technical guide.
Band Combination is used primarily when dealing with multispectral data to allow to use to select a band combination more suited to visualising targets of interest within the imagery. For example with Landsat Imagery the Near Infrared, Red and Green Bands may be used to the digitisation of healthy vegetation.
"Add Buffer to regions" should be a checkbox and not a radio button
This has been Changed
"Add Buffer to regions" and "feature class value" fields should be before output, or are they meant to be changed while editing?
They are after the output box as they can be changed on the fly after the plugin is started.

Tool output:

Difference between output file and existing vector data set not explained.
This has been removed so thee is only one output vector dataset box
Having two separate fields can be confusing. Is one optional? Does one override the other? There could be a dependency, one field becoming active/inactive depending on the other.
This has been removed so thee is only one output vector dataset box
In the layers panel, the added vector layer should have the name given in the "Output vector filename" box, currently has no name.
The vector dataset in the layers menu is now called ‘Digitised Features’
Existing vector layer could give the option to choose from an existing vector data set from the layers panel?
This has been removed so thee is only one output vector dataset box
vector output: why saved with the imagery and no option to choose another location?
This has been removed so thee is only one output vector dataset box
vector output results in a memory layer: would give flexibility to save in other formats, or postgresql
It was decided to commit all changes to a saved vector file after each feature was digitised to reduce the chance work is lost if there is a computer issue. This was observed when testing the plugin operationally in Zanzibar where high temperatures could cause issues to computers while digitisation was being performed.

Tool usability:

Start button should not be active if basic conditions are not met, like having an output file name, or image type.
All conditions must now be met before the start button is activated.
Ideally there should be a toolbar for the red cross circle button, the class and the undo button. Going to the dialogue box next to start and finish for those actions is not intuitive
A toolbar has been added which includes the Undo, Reactivate Tool, Finish, and change feature class number and buffer distance on the fly should the user need this feature.
