### 2021-08-10 #3346 @jorgegil
@gro5-AberUni Thank you for the very comprehensive response to my comments and revision of the paper and plugin.

I'm happy with the changes and recommend the paper to be accepted pending some minor edits.

Under Statement of Need section:
- "Within the open source software domain, there are very region growing tools available." instead of "very" maybe "other" or nothing
- "which are able to select regions within imagery of similar colour," should read "select regions of similar colour within imagery"
- revise "However if desired features are outside of this set distance they will not be included in the vector dataset and a new vector layer will have to be produced to include these layers." Maybe last word should be features?

Under step 4.
- The equation does not need a caption, only a number on the right of it.

- A text revision is recommended for possible typos and style issues.

In the GUI of the plugin:
- Image type: it is still a series of radio buttons without a default checked, when I chose an input image. Maybe you don't want to have the default checked, to make it an explicit choice of the user?
- The 'RGB Drone Image' could be the first item in the radio list.
- The style of the radio button items should be aligned left, like other widgets in the dialog
- Given that there is a toolbar (great addition!) the dialog can disappear once start is clicked and the image appears, to let the user start digitising right away.
