+++ paper/paper.md
@@ -129,7 +129,7 @@ from the selected pixel have a lower weighting. This is done to ensure that pixe
 
 Pixels which can be considered to be of the same region can be defined where the colour distance, summed with the sptial distance have a value greater than the user defined threshold as described in figure 5. Where this condition is not met, pixels are set to set to 0  and considered to be no data values. 
 
-![Equation used to determine if pixels can be considered within the same regino as the users click.](Equation1.png){width="3.937007874015748in"
+![](Equation1.png){width="3.937007874015748in"
 height="3.6587412510936135in"}
 
 **Step 5.**
