+++ paper/paper.md
@@ -30,14 +30,14 @@ bibliography: references.bib
 # Summary
 
 In the fields of image analysis, computer vision and remote sensing
-there are a number of instances where the digitisation of features by
+there are several instances where the digitisation of features by
 the human operator is needed to develop mapped products or for
 generating training data for classification routines. This process is
 notoriously time consuming, labour intensive and error prone and, as
 such, the manual digitisation of features in operational projects is
 often unfeasible, particularly in time-sensitive applications like near
 real-time monitoring applications, emergency response mapping or where
-applications are made at a continental or global scaple. Consequently,
+applications are made at a continental or global scale. Consequently,
 there is a need for a simple to use region growing tool that can be used
 to speed up the digitisation process and minimise human error.
 
@@ -52,7 +52,7 @@ the manual digitising process and minimises error.
 
 # Statement of Need
 
-Within the open source software domain, there are very region growing tools available. Within the Semi-Automatic Classification (SAC) QGIS plugin [@Cogendo2021] is a region growing tool to select like pixels. However, this is intended to be used within the SAC classification workflow and cannot be used intuitively for other digitising needs. In addition there is the ‘Magic Wand’ [@Silvestri2017] plugin which are able to select regions within imagery of similar colour, within a set distance from the point of click and save those to a single vector dataset. However if desired features are outside of this set distance they will not be included in the vector dataset and a new vector layer will have to be produced to include these layers. 
+Within the open-source software domain, there are very region growing tools available. Within the Semi-Automatic Classification (SAC) QGIS plugin [@Cogendo2021] is a region growing tool to select like pixels. However, this is intended to be used within the SAC classification workflow and cannot be used intuitively for other digitising needs. In addition there is the ‘Magic Wand’ [@Silvestri2017] plugin which can select regions within imagery of similar colour, within a set distance from the point of click and save those to a single vector dataset. However if desired features are outside of this set distance, they will not be included in the vector dataset and a new vector layer will have to be produced to include these layers. 
 
 An independent discrete feature region growing tool is needed for two primary reasons: 1) Where users wish to collect training data for conducting image classifications outside of the SAC plugin tool, for instance, taking advantage of the most current and robust machine learning solutions from within Python libraries like TPOT [@olson2016tpot] or Scikit-Learn [@pedregosa2011scikit]. 2) Where users wish to extract thematic information directly rather than following a classification routine. 
 
@@ -62,7 +62,7 @@ The RegionGrow plugin was developed to allow non-experts to identify and digitis
 
 It was then developed to allow for the generation of training datasets which can be used in supervised image classification of a wide range of remotely sensed data, where a single vector dataset is produced. This dataset produces a series of features of differing attributes which relate to the target classes a user is trying to detect. 
 
-The RegionGrow plugin, unlike MagicWand, is not deigned to select all features of similar colour across a whole image as this could introduce false positives where different tarrget features may exhibit different colours. Rather RegionGrow is designed to digitise discrete features identified by the user, where colour is only used to separate that feature from surrounding areas. 
+The RegionGrow plugin, unlike MagicWand, is not deigned to select all features of similar colour across a whole image as this could introduce false positives where different target features may exhibit different colours. Rather RegionGrow is designed to digitise discrete features identified by the user, where colour is only used to separate that feature from surrounding areas. 
 
 # Ongoing Projects
 
[…]
