### 2021-08-11 #3346 @gro5-AberUni
@jorgegil Thank you for your further comments regarding the paper and the plugin GUI, i have updated the paper and made corrections where appropriate.

Regarding the GUI i have adjusted the radio buttons so that the RGB drone button is selected by default when the image file is selected and adjusted the layout of the radio buttons and label so they are consistent with the rest of the items in the dialogue box.
I have also made the change you suggested where the dialogue box is now sent to the background behind the main QGIS window.
