### 2021-08-12 #3346 @Geogoeroe
>
>
> @gro5-AberUni the idea is that you consider @jorgegil comments and improve the plugin and let us know here what was modified and how.
>
> 👋 @Geogoeroe you haven't ticked any of the chechboxes above, is it because you don't have write access? If so I can reinvite you.

Hi all,
I really hate to say and do this, but... I give up.

Today, refreshed from my holiday, I gave it another shot. Everything looks good about this plugin: simple, clean, manual is clear. But I just can't get it to work. Today I downloaded a fresh version of QGIS (latest), fresh install of the newly downloaded plugin, with the extra menu items in the toolbar and all. Python error. Started with a new profile, different Python error. Went back to my LTR 3.16 with a new profile and the new download, python error. Again.

So basically, I've seen it in action one time, before my holiday, there was one moment when I had it running. But after shutting QGIS down and restarting it the next day, it was all errors again (same thing I've been mentioning before, the warning that I should change my parameter settings because it didn't return any pixels from where I clicked. And now I'm back to Python errors again.

It feels like a kind of personal defeat here, especially because this apparently doesn't happen with you guys. Maybe it's windows that's bugging me? Really coudn't say, but... this was a very frustrating experience.

I'm really happy to try again if it works for me, as I am more into trying functionality of plugins, not into debugging errors. Sorry.

Cheers,
Erik.
