### 2021-08-12 #3346 @gro5-AberUni
@Geogoeroe
Hi Erik, I'm sorry to hear that you have been struggling with the plugin, would it be possible to get a screenshot of the python errors that you have received so I can attempt to debug the issues?

Are you receiving an error message saying: 'Not Enough Spectral Difference Found. Increase the Threshold'? If so this error message is no longer present in the most recent version of the plugin, so if you are seeing this message it could perhaps indicate that there is a plugin version conflict somewhere, when you installed the most recent version of the plugin from github did you uninstall the existing version from within the QGIS 'Manage Plugins' window?

Thanks for your feedback,
Greg
