### 2021-08-12 #3346 @Geogoeroe
Hi Greg,
I'll check the error messages tomorrow and will put some here (there were different ones on different occasions, couldn't get a fix on the actual problem, was really weird). For the other error message about the spectral difference: I didn't get that far today, last one was from last month I'd say. Sorry, wrote a little frustrated I think ;-).

Erik.
