### 2021-08-13 #3346 @Geogoeroe
![2021-08-13 12_46_50-_Untitled Project — QGIS  straight from the box!](https://user-images.githubusercontent.com/9982061/129347393-2f40149f-0617-46de-b5cd-ae88d3944beb.png)

This one was the one I could replicate now. Just to be sure, I've really cleaned out my QGIS residual stuff from my computer).

After starting QGIS with a new profile and trying exactly the same, I got this:
![2021-08-13 13_04_00-_Naamloos project — QGIS  brandschoon](https://user-images.githubusercontent.com/9982061/129348028-20ddf794-1e3f-4fa7-9fe5-32ae1e9ed982.png)

Which is interesting, it looks like it cannot create a new geojson file.

I've made a small video (it's unlisted, but [can be reached through this link](http://youtu.be/Nn0GX5ttvuY?hd=1)) to see what's happening here. Maybe I'm really doing something wrong here, the video should show that as well.
