### 2021-08-13 #3346 @hugoledoux
thanks @Geogoeroe for your extra efforts, I hope you still like JOSS 😀

What I propose is this: if @gro5-AberUni can figure out what is wrong then we can continue, if not then as the editor I have to go double-check that users using Windows can install and use it. So I'll try to find someone who uses Windows and is knowledgeable in this area.
