### 2021-08-13 #3346 @Geogoeroe
Tomorrow I'll meet with a few other QGIS friends, I plan to let them check as well, to see it's not just my setup :-)
