### 2021-08-16 #3346 @Geogoeroe
All right, we checked. On 4 different windows laptops the tool is not working. In order to see if it is a windows thing, we also checked on a Linux laptop. Also, not working. Then, after trying a couple of times, it worked, for no apparent reason (on the linux laptop).

My take on this: code is currently too buggy to review properly. Not fit for use yet, sorry to say.
