### 2021-08-16 #3346 @gro5-AberUni
@Geogoeroe I'm sorry to hear you have still had issues, I have checked the errors you have received and the video you uploaded. It seems as though the second issue you're seeing is a result of the incorrect use of the save vector file, following feedback from @jorgegil the plugin GUI and documentation has been updated so that the vector dataset is saved using the system file dialogue box to be in line with other QGIS plugins.
I will continue to debug the first issue you have received as I have not seen this behaviour when I have tested the plugin on windows, Linux or MacOS machines. There is an example dataset included within the plugin folder which I have tested with the plugin and can be used to test functionality
