### 2021-08-16 #3346 @hugoledoux
thanks @Geogoeroe for having the code/plugin tested on other machines.

Unfortunately @gro5-AberUni , to be accepted the code should really run on all 3 major platforms without issues. I propose you indeed try to fix the 2nd issue, and then if you could test it yourself on those platforms that would be best.

The content of the paper and the scope of the work is fine, but we need to have it working.

Let us know when you have fixed the errors.
