### 2021-08-17 #3346 @gro5-AberUni
Hi @Geogoeroe

To assist me in debugging the issue would you be able to let me know of some details regarding the Drone Image you are using so i can try to replicate the error that you are receiving. From digging into the code i think the issue may being caused by some unexpected behaviour with the image itself rather than a platform dependent issue.

Could you please tell me what the image coordinate system is, the image extent and the pixel resolution, knowing this information might help me figure out whats going on with the code.

Thanks very much for your help.

Greg
