### 2021-08-17 #3346 @Geogoeroe
The image is in 32631 - WGS 84 UTM zone 31N. File is attached.

However, the same thing happened with the supplied test data (I tried this with the project CRS was at 4326, at 3857, and (because I wasn't paying attention) with 28992).
[odm_orthophoto.zip](https://github.com/openjournals/joss-reviews/files/6998865/odm_orthophoto.zip)

By the way, I do know about the use of the save vector file. I never do it this way (you always go through the ... button), but the documentation specifically stated that it would automatically save in the same directory as the data would be. But I tried it with the complete path as well.

My other suggestions would be (I would give these in a normal review, but hey):

- you could use the option to load a raster from one of the layers in the project. In general, that makes more sense to end-users than to add another layer (probably one that they ae already viewing).
- delete all text next to the buttons in the toolbar. No other plugin has this, in my experience. It also kind of defeats the purpose of the button. You could make a menu entry instead for textual clarification of the tools.
- geojson and shapefiles are okay as output options, but why not use a geopackage? Makes more sense nowadays.

Cheers,
Erik.
