### 2021-09-13 #3346 @gro5-AberUni
@hugoledoux

I've continued to test the plugin on various different computer systems running Windows, MacOS and Linux operating systems using different combination of QGIS versions on each, and trying a brand new fresh Windows 10 install with QGIs and i have not be been ale to recreate the errors that haver been seen in the comments above. I have used tested using both the example imagery i provide with the plugin, as well as the data provided above by Erik and each combination of system and QGIS version all seemed to work fine with no issues. I've also asked colleagues at other institutions toi test the plugin on their systems and they've reported no issues either.
