### 2021-12-20 #3346 @hugoledoux
I have to conclude that the paper will not be further updated by the author (to fulfill the demands stated above), I have waited more than 1 month for a confirmation now.

We will therefore stop this review, thanks @jorgegil and @Geogoeroe for your contributions.
