### 2021-12-20 #3346 @gro5-AberUni
@hugoledoux Sorry for the long delay in my response, due to time pressures we will not be able to make the necessary adjustments to the plugin to meet the QGIS requirements as outlined above. Thank you all for you contributions and feedback with the plugin.

As I understand it I am able to withdraw the paper through the JOSS portal which I assume will close the review.
